main: move legacy ISO code into legacy_iso.go
This commit moves the rpm based ISO handling from images.go and
main.go into the new legacy_iso.go file to make clear that the
parts in there are only needed for the legacy (rpm) ISO handling
and can be removed once we no longer need to support this.
